streamlit-dataspace-app/requirements.txt
streamlit>=1.30.0
requests>=2.31.0
pandas>=2.0.0
TINTOlib>=1.0.6
pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
bitstring>=4.3.0
